Speed results for a list of URLs

The speedlist function submits a list of URLs to the Google Page Speed Insights API, and returns a dataframe in which each row contains the results from the API for a unique URL in the list.

speedlist(pagelist, strategy, key, filter_third_party_resources = FALSE)

Arguments

pagelist
A list of URLs on which to run the Google Page Speed Insights test.

strategy
Whether the list of URLs should be evaluated in a mobile or desktop context. Accordingly, acceptable values are either the string "mobile" or the string "desktop".

key
A unique key obtainable from Google by registering for free as a Google developer.

filter_third_party_resources
A boolean to indicate if third party resources should be filtered out before PageSpeed analysis. (Default: FALSE)

Examples

# NOT RUN {
speedlist(listofURLs,"mobile",key)
# }